[GH-5202] Not mark objects read only when proxy before.
