+++ lib/Doctrine/ORM/UnitOfWork.php
@@ -2729,10 +2729,10 @@ class UnitOfWork implements PropertyChangedListener
             if ($entity instanceof NotifyPropertyChanged) {
                 $entity->addPropertyChangedListener($this);
             }
-        }
 
-        if (isset($hints[Query::HINT_READ_ONLY])) {
-            $this->readOnlyObjects[$oid] = true;
+            if (isset($hints[Query::HINT_READ_ONLY])) {
+                $this->readOnlyObjects[$oid] = true;
+            }
         }
 
         foreach ($data as $field => $value) {
